Move get_mc_info to mcinfo_io
